TANI/UstSolunumYolu/services/triage_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
pillow>=10.0.0
tensorflow-macos>=2.15.0
tensorflow-metal>=1.1.0
pyyaml>=6.0.1
numpy>=1.24.0
